+++ src/com/palmdigital/rain/entity/mob/Dummy.java
@@ -75,7 +75,7 @@ public class Dummy extends Mob
 	public void render(Screen screen) 
 	{
 		sprite = animSprite.getSprite();
-		screen.renderMob(x - 16, y - 16, sprite, 0);
+		screen.renderMob((int)(x - 16), (int)(y - 16), sprite, 0);
 	}
 	
 }
